refactor: use framework utils and helpers for analytics tests
