+++ e2e/specs/analytics/helpers.ts
@@ -1,4 +1,4 @@
-import { MockttpServer, ServerMockedEndpoint } from 'mockttp';
+import { MockedEndpoint, Mockttp, MockttpServer } from 'mockttp';
 import { E2E_METAMETRICS_TRACK_URL } from '../../../app/util/test/utils';
 
 export interface EventPayload {
@@ -8,19 +8,19 @@ export interface EventPayload {
 
 /**
  * Retrieves payloads of requests matching specified metametrics events.
- * @param {MockttpServer} mockServer - The mock server instance.
+ * @param {MockttpServer|Mockttp} mockServer - The mock server instance.
  * @param {Array<string>} [events] - Event names to filter payloads. If not provided, all events are returned. i.e. ['event1', 'event2']
  * @returns {Promise<Array<EventPayload>>} Filtered request payloads.
  */
 export const getEventsPayloads = async (
-  mockServer: MockttpServer,
+  mockServer: MockttpServer | Mockttp,
   events: string[] = [],
   timeout = 10000,
 ): Promise<EventPayload[]> => {
-  const waitForPendingEndpoints = async (): Promise<ServerMockedEndpoint[]> => {
+  const waitForPendingEndpoints = async (): Promise<MockedEndpoint[]> => {
     const startTime = Date.now();
 
-    const checkPendingEndpoints = async (): Promise<ServerMockedEndpoint[]> => {
+    const checkPendingEndpoints = async (): Promise<MockedEndpoint[]> => {
       const mockedEndpoints = await mockServer.getMockedEndpoints();
 
       // Filter out infrastructure endpoints that are always pending
